Backport minor changes to v3 stable branch (#8097)
* Remove link rake task

All except for one was to do with whitespace. We want to reduce code footprint and use off-the-shelf tools for basic checks like these. Since they aren't critical we are fine without replacements other than what eslint and rubocop already offer.

The sass-rails lint was for webpacker but we have already removed webpacker support. We won't be using any images or asset URLs in the new v4 approach for assets.

* Add metadata to activeadmin.gemspec (#8012)

* Optimize data generation with insert_all

* Mention Ransack allowlist in documentation (#8043)

* Update .gitignore

* Update all outdated dependencies

Ran `bin/bundle update`

Ran `yarn upgrade-interactive --latest`

Note this updates the default gemfile to use Rails 7.1 but we are good with that for our 3-0-stable branch.

---------

Co-authored-by: John DeSilva <desilvjo@umich.edu>
Co-authored-by: Francesco Belladonna <francesco@fc5.me>